requirements.txt
asgiref==3.7.2
Django==4.2.5
django-countries==7.5.1
django-filter==23.1
django-model-utils==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
utils==1.0.1